move generator duties to Agenda
In this version, we moved the responsibility of creating a new
instance of an agenda to Agenda, instead of relying on Generators.

This way, we won't need extra Generator classes.

Upside: fewer classes to deal with, may be easier to comprehend

Downside: uses Class.newInstance. Not recommended because it's
more error prone (see the new Exceptions in the method signature
for markPathToEnd). It should be safe for our purposes though, but
I'll argue the first version has a "cleaner" design